# src/docs/swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
  admin.dashboardRes:
    properties:
      count:
        type: integer
    type: object
  inkstone.ClientError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: yijie.huo@foxmail.com
    name: huoyijie
    url: https://huoyijie.cn
  description: This is ink server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Ink API
  version: "1.0"
paths:
...
  /admin/dashboard:
    get:
      description: Show dashboard
      parameters:
      - description: staff
        enum:
        - staff
        - user
        in: query
        name: category
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.dashboardRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/inkstone.ClientError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/inkstone.ClientError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/inkstone.ClientError'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Show dashboard
      tags:
      - dashboard
...
produces:
- application/json
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"